Risolto il bug che mostrava date errate nei paste salvati localmente. Fixed #14

+++ src/main/java/com/revonline/pastebin/database/PasteDBHelper.java
@@ -113,7 +113,7 @@ public class PasteDBHelper extends SQLiteOpenHelper {
           pasteInfo.setSqlID(cursor.getInt(idxID));
           pasteInfo.setPasteKey(cursor.getString(idxKey));
           Calendar calendar = Calendar.getInstance();
-          calendar.setTimeInMillis(cursor.getInt(idxTime));
+          calendar.setTimeInMillis(cursor.getLong(idxTime));
           pasteInfo.setPasteData(calendar);
 
           list.add(pasteInfo);
